Updated quota usage and limit from int to long to support larger quotas (such as for iCloud)

+++ src/main/java/com/djames42/models/MyQuota.java
@@ -1,20 +1,20 @@
 package com.djames42.models;
 
 public class MyQuota {
-    private int usage;
-    private int limit;
+    private long usage;
+    private long limit;
     private float ratio;
 
-    public MyQuota(int usage, int limit) {
+    public MyQuota(long usage, long limit) {
         this.usage = usage;
         this.limit = limit;
         this.ratio = (float) this.usage / (float) this.limit;
     }
 
-    public int getUsage() {
+    public long getUsage() {
         return usage;
     }
-    public int getLimit() {
+    public long getLimit() {
         return limit;
     }
     public float getRatio() {
